services/honeypot/requirements.txt
psycopg2-binary==2.9.9
redis==5.0.3
structlog==24.1.0
